backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
spotipy==2.23.0
PyMySQL==1.1.0
python-dotenv==1.0.0
marshmallow==3.20.1
email-validator==2.1.0